modularizing the Login ui component, and then reading from jotai atom instead of relying on prop drilling
