+++ src/client/App.tsx
@@ -1,3 +1,4 @@
+import { Provider as JotaiProvider } from "jotai";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { RouterProvider } from "react-router-dom";
 import { MantineProvider } from "@mantine/core";
@@ -11,9 +12,11 @@ const queryClient = new QueryClient();
 export const App = () => {
   return (
     <MantineProvider>
-      <QueryClientProvider client={queryClient}>
-        <RouterProvider router={router} />
-      </QueryClientProvider>
+      <JotaiProvider>
+        <QueryClientProvider client={queryClient}>
+          <RouterProvider router={router} />
+        </QueryClientProvider>
+      </JotaiProvider>
     </MantineProvider>
   );
 };
